+++ Docs/API/html/search/all_4.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['findfunction',['FindFunction',['../class_jinx_1_1_i_script.html#a5bfae4939961acf038b0c34fb676ba69',1,'Jinx::IScript']]],
+  ['findfunction',['FindFunction',['../class_jinx_1_1_i_script.html#afbd7cee1f07529610818524402933b8d',1,'Jinx::IScript']]],
   ['freefn',['freeFn',['../struct_jinx_1_1_global_params.html#a82b02db5c24487a99d0c0ee83cb66c53',1,'Jinx::GlobalParams::freeFn()'],['../_jinx_8h.html#a466c508de040a6f6ef5873b2793e92eb',1,'Jinx::FreeFn()']]]
 ];